# frontend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "transac",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "List products by store ID",
        "operationId": "list_products",
        "parameters": [
          {
            "name": "store_id",
            "in": "query",
            "description": "Store ID to filter products",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Products found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductModel"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid store ID"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Create a new product",
        "operationId": "create_product",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Product created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid data"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "summary": "Get a product by ID",
        "operationId": "get_product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductModel"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Update a product by ID",
        "operationId": "update_product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Product updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid data"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete a product by ID",
        "operationId": "delete_product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted successfully"
          },
          "400": {
            "description": "Bad request - invalid data"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/products/{id}/media": {
      "post": {
        "tags": [
          "Products"
        ],
        "summary": "Upload media for a product",
        "operationId": "upload_product_media",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Media uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid image or analysis failed"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error - upload failed"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "summary": "Replace media for a product",
        "operationId": "edit_product_media",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "text/plain": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Media replaced successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid image or analysis failed"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error - upload failed"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "summary": "Delete media for a product",
        "operationId": "delete_product_media",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Media deleted successfully"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal server error - deletion failed"
          }
        }
      }
    },
    "/products/{id}/reviews": {
      "get": {
        "tags": [
          "Reviews"
        ],
        "summary": "List reviews for a product",
        "operationId": "list_reviews",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Product ID",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reviews found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "post": {
        "tags": [
          "Reviews"
        ],
        "summary": "Create a new review for a product",
        "operationId": "create_review",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Review created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewModel"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid data"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateProductRequest": {
        "type": "object",
        "required": [
          "name",
          "image_ids",
          "price",
          "quantity_available",
          "category",
          "return_policy"
        ],
        "properties": {
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "image_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "quantity_available": {
            "type": "integer",
            "format": "int32"
          },
          "return_policy": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "MediaUploadResponse": {
        "type": "object",
        "required": [
          "image_id",
          "s3_key"
        ],
        "properties": {
          "image_id": {
            "type": "string",
            "format": "uuid"
          },
          "s3_key": {
            "type": "string"
          }
        }
      },
...
      "UpdateProductRequest": {
        "type": "object",
        "required": [
          "name",
          "image_ids",
          "price",
          "quantity_available",
          "category",
          "return_policy"
        ],
        "properties": {
          "category": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "image_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "quantity_available": {
            "type": "integer",
            "format": "int32"
          },
          "return_policy": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}